+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 #Changelog
 
+## 2016-11-21
+- `$.promisify` wrapper value and return promise object
+- `$.asyncEach` Applies the function iteratee to each item in coll
+
 ## 2016-07-27
 - `$.series` Add pure function support
 - `$.series` Add `initialValue` param. when deferreds is array, initialValue is the first deferred function param